% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing-CE.R
\name{read_and_analyze_form_CE}
\alias{read_and_analyze_form_CE}
\title{Reads a form 3-CE (intermediate format) and returns the analysis results for both the METADATA and DATA sections}
\usage{
read_and_analyze_form_CE(filename)
}
\description{
Reads a form 3-CE (intermediate format) and returns the analysis results for both the METADATA and DATA sections
}
